// backoffice/src/app/pages/interfaces/Istore.ts
export interface Store {
  id: string;
  name: string;
  logo: string;
  email: string;
  phone_number_1: string;
  phone_number_2: string;
  description: string;
  location: string;
  facebook_link: string;
  whatsapp_link: string;
  instagram_link: string;
  twitter_link: string;
  is_active: boolean;
  can_evaluate: boolean;
  date_added: string;
  slug: string;
  subscription: string;
}